requirements.txt
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
Cython==0.29.26
Flask==2.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mypy==0.931
mypy-extensions==0.4.3
numpy==1.22.0
pathspec==0.9.0
Pillow==9.0.0
platformdirs==2.4.1
Pyfhel==2.3.1
requests==2.27.1
tomli==1.2.3
typing-extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.2
